Moved ControlBar.tsx to its own folder
