+++ src/components/academy/grading/GradingWorkspace.tsx
@@ -18,7 +18,7 @@ import {
 } from '../../assessment/assessmentShape';
 import Markdown from '../../commons/Markdown';
 import Workspace, { WorkspaceProps } from '../../workspace';
-import { ControlBarProps } from '../../workspace/ControlBar';
+import { ControlBarProps } from '../../workspace/controlBar/ControlBar';
 import {
   ClearButton,
   EvalButton,
